Fixes #16: Detecting EOF for serial access
The program was originally looking for the EOF marker, but I forgot that when the EOF marker BGZF block is decompressed, it evaluates to an empty bytestring `b''`. I put in some extra checks so that if an empty block is found, it ensures it is at the file and is a legitimate end of file.

If the file is truncated, it will `raise` a slightly different `StopIteration` error stating that the *potential* end of file is reached.

I ran this on the file you mentioned in #15 (btw, awesome Issue suggestion. Thank you for following the template). My debugging is properly running through the file now.