Fixing pathing for `version` file and `example_bam`
Lots of going back and forth with fixing how certain resources are packaged in `bamnostic`. Changes to `setuptools` deprecated `pkg_resources` causing errors on newer builds and changes to PEPs are requiring certain additions to `pyproject.toml`. This, coupled with migrations from `setup.py` have made this somewhat of a headache when continuing support for 2.7.